requirements.txt
python-dotenv
moviepy
undetected-chromedriver
selenium-stealth